+++ bridges/compat/workflow_test_bridge.py
@@ -44,8 +44,8 @@ from bridges.common.bootstrap import setup_environment
 setup_environment()
 
 from bridges.common.ipc import IPC
-from bridges.common.connection import ConnectionService
-from bridges.common.app_manager import AppService
+from bridges.common.device.connection import ConnectionService
+from bridges.common.device.app_manager import AppService
 from loguru import logger
 from taktik.core.compat.selectors.tracer import SelectorTracer