+++ bridges/instagram/account_bridge.py
@@ -46,8 +46,8 @@ from bridges.instagram.base import (
     logger, _ipc,
     send_message, send_status, send_error, send_log,
 )
-from bridges.common.connection import ConnectionService
-from bridges.common.app_manager import AppService
+from bridges.common.device.connection import ConnectionService
+from bridges.common.device.app_manager import AppService
 from bridges.common.signal_handler import setup_signal_handlers